Update: Renamve files to match project file name convention + fix ipc type
