+++ src/ipc-handlers.ts
@@ -19,7 +19,7 @@ import { downloadSqliteIntegrationPlugin } from '../vendor/wp-now/src/download';
 import { LIMIT_ARCHIVE_SIZE } from './constants';
 import { isEmptyDir, pathExists, isWordPressDirectory, sanitizeFolderName } from './lib/fs-utils';
 import { getImageData } from './lib/get-image-data';
-import { importBackup } from './lib/import-export/import/importManager';
+import { importBackup } from './lib/import-export/import/import-manager';
 import { allImporters } from './lib/import-export/import/importers';
 import { DbConfig, BackupArchiveInfo } from './lib/import-export/import/types';
 import { allValidators } from './lib/import-export/import/validators';
@@ -101,6 +101,22 @@ export async function getInstalledApps( _event: IpcMainInvokeEvent ): Promise< I
 		vscode: isInstalled( 'vscode' ),
 		phpstorm: isInstalled( 'phpstorm' ),
 	};
+}
+
+export async function importSite(
+	_event: IpcMainInvokeEvent,
+	{ id, backupFile }: { id: string; backupFile: BackupArchiveInfo }
+) {
+	const site = SiteServer.get( id );
+	if ( ! site ) {
+		throw new Error( 'Site not found.' );
+	}
+	const sitePath = site.details.path;
+	try {
+		await importBackup( backupFile, sitePath, {} as DbConfig, allValidators, allImporters );
+	} catch ( e ) {
+		console.log( 'Error importing site', ( e as Error ).message );
+	}
 }
 
 // Use sqlite database and db.php file in situ
@@ -677,21 +693,4 @@ export async function promptWindowsSpeedUpSites(
 	{ skipIfAlreadyPrompted }: { skipIfAlreadyPrompted: boolean }
 ) {
 	await windowsHelpers.promptWindowsSpeedUpSites( { skipIfAlreadyPrompted } );
-}
-
-export async function importSite(
-	_event: IpcMainInvokeEvent,
-	id: string,
-	file: BackupArchiveInfo
-) {
-	const site = SiteServer.get( id );
-	if ( ! site ) {
-		throw new Error( 'Site not found.' );
-	}
-	const sitePath = site.details.path;
-	try {
-		await importBackup( file, sitePath, {} as DbConfig, allValidators, allImporters );
-	} catch ( e ) {
-		console.log( 'Error importing site', ( e as Error ).message );
-	}
 }
