+++ src/lib/import-export/import/importers/index.ts
@@ -1,9 +1,9 @@
 import { BackupContents } from '../types';
 import { Importer } from './Importer';
-import { JetpackImporter } from './JetpackImporter';
+import { JetpackImporter } from './jetpack-importer';
 
 export * from './Importer';
-export * from './JetpackImporter';
+export * from './jetpack-importer';
 
 export const allImporters: { [ key: string ]: new ( contents: BackupContents ) => Importer } = {
 	JetpackValidator: JetpackImporter,
