+++ src/lib/import-export/import/validators/index.ts
@@ -1,7 +1,7 @@
-import { JetpackValidator } from './JetpackValidator';
+import { JetpackValidator } from './jetpack-validator';
 import { Validator } from './Validator';
 
 export * from './Validator';
-export * from './JetpackValidator';
+export * from './jetpack-validator';
 
 export const allValidators: Validator[] = [ new JetpackValidator() ];
