+++ src/lib/tests/import-export/import/importer/jetpack-importer.test.ts
@@ -1,6 +1,6 @@
 // To run tests, execute `npm run test -- src/lib/tests/import-export/import/importer/jetpack-importer.test.ts``
 import * as fs from 'fs/promises';
-import { JetpackImporter } from '../../../../import-export/import/importers/JetpackImporter';
+import { JetpackImporter } from '../../../../import-export/import/importers/jetpack-importer';
 import { BackupContents, DbConfig } from '../../../../import-export/import/types';
 
 jest.mock( 'fs/promises' );
